feat(postcss): Validate format of postcss loader options.
**Why:** An initial stab at how validation should work (?): Steering the user
towards canonical use of the loader in question by prompting him/her to match
the format of the examples shown in the loader's README.

This clears up user's confusion wether the format in which they specified it is still
valid. The specific problem here was that I had specified the option
values of postcss-plugin as an array (`postcss: [...]`) while the
documentation showed them only as `postss: function() { return [...] }`.
It was thus unclear to me wether some major update of the loader would
break my build or if my format was still ok. Only by looking into the
loaders source could I see that my format would still be ok.